Updated Linux pathing
Updated the Linux path for all files to use the absolute path. This also means that the repository must be cloned to the desktop of the Virtual Machine in order to work.
Currently working on trying to get the system to work on MAC. Linux and Windows now work fully.